Refactor packages to com.example namespace
