+++ src/Miner/src/com/example/dreambot/miner/powerminer/gui/ScriptVars.java
@@ -1,0 +1,12 @@
+package com.example.dreambot.miner.powerminer.gui;
+
+import java.util.ArrayList;
+import java.util.List;
+import com.example.dreambot.miner.scriptmain.powerminer.MineTask;
+
+
+public class ScriptVars {
+
+	public List<MineTask> tasks = new ArrayList<>();
+	public boolean started = false;
+}
