--- src/Miner/src/nezz/dreambot/powerminer/gui/ScriptVars.java
@@ -1,12 +1,0 @@
-package nezz.dreambot.powerminer.gui;
-
-import java.util.ArrayList;
-import java.util.List;
-import nezz.dreambot.scriptmain.powerminer.MineTask;
-
-
-public class ScriptVars {
-
-	public List<MineTask> tasks = new ArrayList<>();
-	public boolean started = false;
-}
