--- src/Walker/src/GUI/Window.form
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="GUI.Window">
-  <grid id="27dc6" binding="RootPanel" layout-manager="GridLayoutManager" row-count="1" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="55" y="71" width="585" height="349"/>
-    </constraints>
-    <properties>
-      <font name="Open Sans"/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <tabbedpane id="bf078" binding="tabs">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false">
-            <preferred-size width="200" height="200"/>
-          </grid>
-        </constraints>
-        <properties/>
-        <border type="none"/>
-        <children>
-          <grid id="6166b" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-            <margin top="0" left="0" bottom="0" right="0"/>
-            <constraints>
-              <tabbedpane title="Walk location"/>
-            </constraints>
-            <properties/>
-            <border type="none" title="Walk Location"/>
-            <children>
-              <component id="f474d" class="javax.swing.JLabel">
-                <constraints>
-                  <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties>
-                  <text value="Location"/>
-                </properties>
-              </component>
-              <vspacer id="affb5">
-                <constraints>
-                  <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
-                </constraints>
-              </vspacer>
-              <component id="42f91" class="javax.swing.JComboBox" binding="walkLocation">
-                <constraints>
-                  <grid row="1" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties>
-                  <model>
-                    <item value="Al Kharid Bank"/>
-                    <item value="Al Kharid Mining spot"/>
-                    <item value="Barbarian village"/>
-                    <item value="Chaos Temple"/>
-                    <item value="Draynor"/>
-                    <item value="Duel Arena"/>
-                    <item value="Edgeville Monastery"/>
-                    <item value="Edgeville Bank"/>
-                    <item value="Falador East"/>
-                    <item value="Falador West"/>
-                    <item value="G.E."/>
-                    <item value="Hill Giants Dungeon"/>
-                    <item value="Lumbridge"/>
-                    <item value="Port Sarim"/>
-                    <item value="Rimmington"/>
-                    <item value="Varrock Center"/>
-                    <item value="Varrock East"/>
-                    <item value="Varrock West"/>
-                    <item value="Wizard's Tower"/>
-                  </model>
-                </properties>
-              </component>
-              <component id="9d75e" class="javax.swing.JButton" binding="startButton" default-binding="true">
-                <constraints>
-                  <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties>
-                  <text value="Start!"/>
-                </properties>
-              </component>
-            </children>
-          </grid>
-        </children>
-      </tabbedpane>
-    </children>
-  </grid>
-</form>
